Ll/bug fixes

See merge request cpsc319-2018w2/jabc/oecm/jabc-hr!93

+++ src/service/PerformanceService.ts
@@ -68,6 +68,10 @@ export async function deleteComment(id: Number, idComment: Number, xAuthToken: s
 export async function deletePerformancePlan(id: Number, xAuthToken: string) {
     try {
         const performancePlan = await getPerformancePlan(id, xAuthToken)
+        const client = (await Auth(xAuthToken)).employee
+        if (client.adminLevel === IEmployee.adminLevelEnum.STAFF) {
+            throw new JABCError(JABCResponse.EMPLOYEE, 'An employee with STAFF admin level can not delete performance plan');
+        }
         await Database.getInstance().query('CALL delete_performance_plan(?)', [id], JABCResponse.PERFORMANCE)
 		return new JABCSuccess(JABCResponse.EMPLOYEE, 'The performance plan was successfully deleted')
 	} catch (error) {
@@ -90,7 +94,6 @@ export async function deletePerformanceReview(id: Number, xAuthToken: string) {
         const client = (await Auth(xAuthToken)).employee
         if (client.id == performanceReview.fkEmployee && client.adminLevel != IEmployee.adminLevelEnum.HR_ADMIN) {
             throw new JABCError(JABCResponse.EMPLOYEE, 'Can not delete your own performance review')
-
         }
 		await Database.getInstance().query('CALL delete_performance_review(?)', [id], JABCResponse.PERFORMANCE)
 		return new JABCSuccess(JABCResponse.EMPLOYEE, 'The performance review was successfully deleted')
@@ -152,10 +155,12 @@ export async function getPerformancePlan(id: Number, xAuthToken: string) {
         const client = (await Auth(xAuthToken)).employee
         let res = await Database.getInstance().query('CALL get_performance_plan(?)', [id], JABCResponse.PERFORMANCE)
         let performancePlan = new PerformancePlan(res[0][0][0])
-        if (client.adminLevel == IEmployee.adminLevelEnum.MANAGER) {
-            await isManagedBy(performancePlan.fkEmployee, client.id)
-		}else if(client.adminLevel == IEmployee.adminLevelEnum.STAFF && client.id != performancePlan.fkEmployee){
-            throw new JABCError(JABCResponse.EMPLOYEE, 'An employee with STAFF admin level can not manage other employee\'s performance plan.')
+        if (client.id !== performancePlan.fkEmployee) {
+            if (client.adminLevel == IEmployee.adminLevelEnum.MANAGER) {
+                await isManagedBy(performancePlan.fkEmployee, client.id)
+            } else if (client.adminLevel == IEmployee.adminLevelEnum.STAFF) {
+                throw new JABCError(JABCResponse.EMPLOYEE, 'An employee with STAFF admin level can not manage other employee\'s performance plan.')
+            }
         }
         performancePlan.sections = PerformanceSection.PerformanceSections(res[0][1])
         performancePlan.comments = Comment.Comments(res[0][2])
@@ -179,13 +184,16 @@ export async function getPerformanceReview(id: Number, xAuthToken: string) {
         const client = (await Auth(xAuthToken)).employee
         let res = await Database.getInstance().query('CALL get_performance_review(?)', [id], JABCResponse.PERFORMANCE)
         let performanceReview = new PerformanceReview(res[0][0][0])
-        if (client.adminLevel == IEmployee.adminLevelEnum.MANAGER) {
+        if (client.adminLevel == IEmployee.adminLevelEnum.MANAGER && client.id !== performanceReview.fkEmployee) {
 			await isManagedBy(performanceReview.fkEmployee, client.id)
 		}else if(client.adminLevel == IEmployee.adminLevelEnum.STAFF && client.id != performanceReview.fkEmployee){
             throw new JABCError(JABCResponse.EMPLOYEE, 'An employee with STAFF admin level can not manage other employee\'s performance review.')
-        } else if (performanceReview.status == 0 && performanceReview.fkEmployee === client.id && client.adminLevel !== IEmployee.adminLevelEnum.HR_ADMIN) {
+        }
+
+        if (performanceReview.status == 0 && performanceReview.fkEmployee === client.id && client.adminLevel !== IEmployee.adminLevelEnum.HR_ADMIN) {
             throw new JABCError(JABCResponse.EMPLOYEE,'An employee can not see its own performance review when unpublished' )
         }
+
         performanceReview.sections = PerformanceSection.PerformanceSections(res[0][1])
         performanceReview.comments = Comment.Comments(res[0][2])
         return performanceReview
